Add some basic tests
- uses mocha, chai, supertest to run some basic tests
- could extend further using sinon to mock the JIRA API
- separates the actual express app from the startup script (index.js) so
  that trying to run the tests don't start the real server
