+++ .github/workflows/manual-deploy-testing.yaml
@@ -1,0 +1,45 @@
+---
+name: Manual Deploy to Testing
+
+on:
+  workflow_dispatch:
+    inputs:
+      tag:
+        description: tag
+        required: true
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      packages: write
+
+    steps:
+      - uses: hathitrust/github_actions/build@v1
+        with:
+          image: ghcr.io/${{ github.repository }}-unstable
+          tag: ${{ github.event.inputs.tag }}
+          push_latest: false
+          registry_token: ${{ github.token }}
+
+  deploy:
+    needs: build
+    runs-on: ubuntu-latest
+    environment: testing
+    steps:
+      - uses: actions/checkout@v3
+      - name: Find commit for tag
+        id: tag_check
+        uses: hathitrust/github_actions/validate-tag@v1
+        with:
+          tag: ${{ github.event.inputs.tag }}
+      - name: Deploy to testing
+        uses: mlibrary/deploy-to-kubernetes@v3
+        with:
+          registry_token: ${{ github.token }}
+          image: ghcr.io/${{ github.repository }}-unstable:${{ steps.tag_check.outputs.tag }}
+          cluster_ca: ${{ secrets.KUBERNETES_CA }}
+          cluster_server: ${{ secrets.KUBERNETES_SERVER }}
+          namespace_token: ${{ secrets.KUBERNETES_TOKEN }}
+          namespace: ${{ secrets.KUBERNETES_NAMESPACE }}
