+++ .github/workflows/tests.yaml
@@ -1,0 +1,24 @@
+name: Run Tests
+
+on:
+  push:
+    branches:
+      - main
+
+  pull_request:
+
+jobs:
+  # Run tests
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Set up tests
+        run: |
+          docker compose build
+          docker compose run test npm install
+
+      - name: Run tests
+        run: docker compose run test
