+++ docker-compose.yml
@@ -1,0 +1,24 @@
+version: '3'
+
+services:
+  web:
+    build:
+      context: .
+      target: development
+    ports:
+      - "5000:5000"
+    volumes:
+      - .:/usr/src/app
+      - node_cache:/usr/src/app/node_modules
+
+  test:
+    build:
+      context: .
+      target: development
+    command: npm test
+    volumes:
+      - .:/usr/src/app
+      - node_cache:/usr/src/app/node_modules
+
+volumes:
+  node_cache:
